Fix: Explicit Capture by Copy in streaming_test.cc for MSVC 2019
We set an explicit default of capture-by-copy (i.e. by value) for
certain lambda functions, and then specify the parameters that are to
be captured by reference. In MSVC 2019, even for const/constexpr
implicit lambda-function parameters, the compiler requires that the
default capture method be provided when implicitly passing
const/constexpr parameter(s) to lambda function by copy/value.