Fix: Avoid using compiler.has_function('dladdr')
It would be nice to use compiler.has_function('dladdr') to test if we
need to add libdl, but unfortunately that approach seems to break
certain Redhat builds, specifically the Redhat builds that use
ubi-8. In such cases, it appears meson creates a test file that
includes the comment:
  With some toolchains ... the compiler provides various builtins
  which are not really implemented... [If] the user provides a header,
  including the header didn't lead to the function being defined, and
  the function we are checking isn't a builtin itself, we assume the
  builtin is not functional and error out
To avoid generating such an error, we take the simpler approach of
trying to add libdl but making it optional (i.e. not required).

Also - added branch MakeDlLibraryDependency4 to linux.yaml as a
temporary experiment.