Fix: Avoid Doing "brew update" for macOS Runner
We avoid doing "brew update" - the brew formulas that are preinstalled
on the github runner image are likely consistent with the
pre-installed software on the image. If we do "brew upate", and then
install something new with brew (specifically meson), and the
"something new" depends on pre-installed software on the image, and
there are new versions of the pre-installed software revealed by doing
"brew update", then when we install the "something new" brew may try
and also install a new version of the pre-installed software on which
the "something new" depends, but that attempt to install a new version
of the pre-installed software can fail as a result of being blocked by
the software that is already installed.

Also, removed this branch from list of branches to run macos.yaml.