Fix: Install latest libstdc++ in workflow for debian:testing/clang
Intermittently, debian:testing fails with clang, saying that libstdc++
cannot be found. In debian:testing/clang, normally libstdc++-dev is
installed as a dependency of meson. In the situation where the build
breaks, the meson dependency is different to the latest available
libstdc++; for instance, installing meson dependencies might install
libstdc++-13-dev, whereas apt has access to libstdc++-14-dev. In other
situations, the meson dependency libstdc++-..-dev may be the same as
the latest libstdc++-..-dev to which apt has access.

To prevent the build breaking, we need to install the latest
libstdc++-..-dev to which apt has access - not the N-1 version which
may be installed as a meson dependency. We do this now in linux.yaml.