Fix: Dynamic Sleep Time in rest_server_test; Update windows.yaml
In rest_server_test / keepalive_multithread_client_request, it sleeps
for a certain period before testing that a set of peers have been
successfully connected. The client used to connect the peers is
httplib.h. For Linux, macOS, and BSD, and sleep time of 700ms works
well. However, in Windows there are significant variaions in the ideal
sleep time depending on the version of the OS. Accordingly, we now
determine the total sleep time dynamically in 700ms increments.

Note: This build runs all tests successfully on Windows Server 2019
with Visual Studio 2019, and on Windows 11 with Visual Studio 2022,
for both debug and non-debug builds.

Also: Update windows.yaml