Fix: Avoid Making Most Windows Libraries Statically Linked
Previously, when building with gcc, we had made most of the required
Windows libraries (Ws2_32, Iphlpapi etc.) link statically with
pistache.dll. This turns out to be unnecessary given other fixes, and
we have gone back to linking those libraries dynamically.

libgcc and libstdc++ really do need to be statically linked with the
executable in gcc builds, and we do that in tests/meson.build and
examples/meson.build.

Also: Update windows.yaml