Fix: Load SetThreadDescription Dynamically for mingw
When in mingw and in server versions of Windows, the Windows API
function SetThreadDescription seems to be missing from the header
file. So for mingw, we link to SetThreadDescription dynamically / at
run time in reactor.cc (where it is used).

Also: windows.yaml updates

Also: Outputting actual_fd in listener

Doing PS_LOG_INFO in listener for the actual_fd that will be used as
an index into the array of Transports. Curious if it is still a
multiple of 4 in Windows with gcc.