Fix: Use Incrementing Counter to derive Handlers Index in Windows
In listener.cc, for linux/macOS/BSD, we use the peer's actual-fd as an
input to determine which handler/Transport to use to service each
peer, thus distributing the peers amongst the Listener's set of
handlers. However, in Windows actual_fd seems to be a multiple of
4, so we'll fail to use a bunch of handlers if we just do "idx =
actual_fd % handlers.size()". For instance, if handlers.size() is 4,
idx will always be zero - and this actually happens in the test
http_client_test.test_client_timeout.

We had been using the crude fix in Windows of simply dividing the
actual-fd by 4. However, this fails (i.e. does not cause the peers to
be well distrbuted amongst the handlers) intermittently in Windows
Server 2019, compiling Pistache with gcc. So now we fix it "properly'
for Windows by using an incrementing counter, belonging to the
Listener, to generate the input to determine the handler index to use
for each newly-attached peer.

Also: Update windows.yaml