Fix: Correct "meson install" issue in Windows-gcc case
installman.ps1
installmanatinstall.ps1
When using gcc, the linker forces pistachelog to have a 'lib' filename prefix.
On install, we copy that filename to "pistachelog.dll", so the
manifest install process can access it.

meson.build
Use the correct form of the name of the static pistachelog library for gcc.

winscripts/commonsetup.ps1
Corrected a small issue with the way git was being installed

Corrected Some windows.yaml Formatting