Fix: Use ByPass ExecutionPolicy when Invoking installman.ps1
installman.ps1 installs the logging manifest as part of the build
process on Windows. By explicitly using a Windows ExecutionPolicy of
ByPass in meson.build when invoking installman.ps1, we allow build to
complete successfully even when the default ExecutionPolicy is
Restricted (an ExecutionPolicy of Restricted prevents ps1 scripts from
executing).