fix: don't break C++23's std::unreachable
If "winornix.h" were to be included before <utility> then the standard
definition of std::unreachable is broken because the 'unreachable' is
replaced by the Pistache macro content.
This patch avoids using a macro, puts the Pistache implementation in a
detail namespace and simply pulls in the standard function if available.