meson: check for availability of dladdr
Delete erroneous comment claiming that compiler.has_function is broken.
This is based on a misunderstanding of the original logfile.
See: https://github.com/pistacheio/pistache/pull/1252#discussion_r1924598159

Checking for dladdr works fine. Checking for `dependency('dl')` on
versions of meson that are too old to support it does not work, and this
is what happened on RedHat ubi-8.