CI: Add Check That Windows Logging Is Working
windows.yaml: Now runs a test with logging enabled, and checks that a)
An INFO message has been correctly wrriten to the log; and b) i) for
debug builds, checks that a DEBUG message is written to the log, or
ii) for non-debug builds, checks that a DEBUG message is NOT written
to the log.

pist_sockfns.cc: Make sure that at least one INFO message is written
to the log (it is written on winsock initialization).

Also - if winsock initialization fails, log a WARNING, not just a
DEBUG message.