issue-1312: more robust Listener::shutdown()
The shutdownFd is used to tell the acceptor worker thread to exit,
when there is multiple threads, the current implementation is not
enough to ensure every worker has a chance to get the event, if it's
busy doing something, it will miss the notification.

Without changing the NotifyFd interface, the guaranteed way is to
do a loop notification and check the number of alive worker threads,
keep notifying them until all the threads are done.